pyproject.toml
[tool.poetry]
name = "simpplax"
version = "0.1.0"
description = ""
authors = ["McCoy R. Becker <mccoyb@mit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
jax = "^0.4.19"
rich = "^13.6.0"
beartype = "^0.16.4"
jaxtyping = "^0.2.23"
plum-dispatch = "^2.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
pytest = "^7.2.0"
coverage = "^7.0.0"
pytest-benchmark = "^4.0.0"
pytest-xdist = {version = "^3.2.0", extras = ["psutil"] }
xdoctest = "^1.1.0"
safety = "^2.3.5"
jupyterlab = "^3.5.1"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
# Enable the isort rules.
extend-select = ["I"]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

